Update package.json
Adding gatsby as a dependency could lead to strange behavior when the current gatsby version is lower than what specified in the plugin. This will also give issues when a new gatsby major comes out.

Removed react as dependencies as well as I don't think it's used. If so, add them to peerdeps for the same reason as above.